Add tests for load_account
